[Refactor] 코드 리팩토링4
