+++ contents/todoListAPI/kangwook/practice/src/main/java/com/appcenter/practice/PracticeApplication.java
@@ -10,5 +10,4 @@ public class PracticeApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(PracticeApplication.class, args);
 	}
-
 }
