+++ contents/todoListAPI/kangwook/practice/src/main/java/com/appcenter/practice/domain/Member.java
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -41,23 +40,17 @@ public class Member extends BaseEntity implements UserDetails {
     private List<Todo> todoList =new ArrayList<>();
 
     @Builder
-    private Member(String email, String password, String nickname) {
+    private Member(String email, String password, String nickname, Role role) {
         this.email = email;
         this.password = password;
         this.nickname = nickname;
+        this.role= role;
     }
 
     public void changePassword(String password){
         this.password=password;
     }
     public void changeNickname(String nickname){ this.nickname=nickname;}
-    public void addUserAuthority(){
-        this.role=Role.ROLE_USER;
-    }
-    public void passwordEncode(PasswordEncoder passwordEncoder){
-        this.password=passwordEncoder.encode(password);
-    }
-
 
     //grantedAuthority는 부여된 권한을 갖는 인터페이스,SimpleGrantedAuthority는 granteAuthority를 구현한 간단한 클래스
     //"ROLE_USER", "ROLE_ADMIN"과 같은 권한을 나타내는 문자열을 SimpleGrantedAuthority 객체로 생성하여 사용할 수 있다.
