+++ contents/todoListAPI/kangwook/practice/src/main/java/com/appcenter/practice/dto/request/comment/UpdateCommentReq.java
@@ -1,4 +1,4 @@
-package com.appcenter.practice.dto.reqeust.comment;
+package com.appcenter.practice.dto.request.comment;
 
 
 import jakarta.validation.constraints.NotBlank;
