+++ contents/todoListAPI/kangwook/practice/src/main/java/com/appcenter/practice/dto/request/todo/AddTodoReq.java
@@ -1,4 +1,4 @@
-package com.appcenter.practice.dto.reqeust.todo;
+package com.appcenter.practice.dto.request.todo;
 
 import com.appcenter.practice.domain.Member;
 import com.appcenter.practice.domain.Todo;
