+++ contents/todoListAPI/kangwook/practice/src/main/java/com/appcenter/practice/service/TodoService.java
@@ -3,8 +3,8 @@ package com.appcenter.practice.service;
 
 import com.appcenter.practice.domain.Member;
 import com.appcenter.practice.domain.Todo;
-import com.appcenter.practice.dto.reqeust.todo.AddTodoReq;
-import com.appcenter.practice.dto.reqeust.todo.UpdateTodoReq;
+import com.appcenter.practice.dto.request.todo.AddTodoReq;
+import com.appcenter.practice.dto.request.todo.UpdateTodoReq;
 import com.appcenter.practice.dto.response.todo.ReadTodoRes;
 import com.appcenter.practice.exception.CustomException;
 import com.appcenter.practice.common.StatusCode;
